+++ NEWS.md
@@ -1,11 +1,12 @@
 fasstr 0.3.2
 =========================
 
-Updated: 13 Jan 2020
+Updated: 20 Jan 2020
 
 #### Updates:
 
-  * README updated: sticker, badges
+  * Added calc_annual_peaks() function to calculate both annual n-day min. and max. flows with dates of occurrence.
+  * README updated: sticker, badges.